+++ CHANGELOG.md
@@ -2,6 +2,55 @@
 
 All notable changes to the "launchdarkly" extension will be documented in this file.
 
+
+## [6.2.0] - 2026-01-28
+
+### Added
+
+- Added "EU" as a first-class option in the instance picker during sign-in, making it easier for EU customers to connect to `app.eu.launchdarkly.com`
+
+## [6.1.0] - 2025-12-05
+- Bumping major and minor version to address rogue published version that is not captured in our package.json
+
+
+## [5.1.0] - 2025-09-01
+ 
+### Fixed
+
+- Bug with token validation
+- Updated broken quick link urls 
+- Upgraded failing dependencies 
+
+### Changed
+
+- Login flow
+- Better handling of invalid sessions 
+
+### Added 
+
+- Ability to sign out of session
+- Custom error handling for failed authentication 
+- Developement instructions into README
+
+## [5.0.0] - 2024-02-06
+
+### Fixed
+
+- Numerous performance issues related to code lens.
+- Initial setup now recovers better if there's been a problem.
+
+## Changed
+- Initial setup is now split into Sign in and Configuration.
+- Updated CLI version of Code References.
+
+## Added
+- Code lens and hover now show SDK availability of a flag.
+- Sign via through AuthProvider API.
+- `LaunchDarkly: Flag Actions` command.
+- Quick Targeting allows you to add/remove single context or rule from a flag.
+- Flags in File now has inline commands on entry.
+- Status bar icon that tells you current configured project and environment.
+
 ## [4.0.2] - 2023-02-15
 
 ### Fixed